set: 파일구조 변경 후 블로그 게시물 생성 구현 및 테스트 코드 작성 issue: SCRUM-123
