--- src/modules/blogs/interfaces/applications/create.blog.application.interface.ts
@@ -1,6 +1,0 @@
-import { CreateBlogDomain } from '../../domain/request/create.blog.domain';
-import { BlogEntity } from '../../domain/blog.entity';
-
-export interface CreateBlogApplication {
-    createBlog(blogDomain: CreateBlogDomain): Promise<BlogEntity>;
-}
