--- src/modules/blogs/interfaces/applications/get.blog.application.interface.ts
@@ -1,5 +1,0 @@
-import { GetBlogDomain } from '../../domain/response/get.blog.domain';
-
-export interface GetBlogApplication {
-    getBlog(blogId: number): Promise<GetBlogDomain>;
-}
