--- src/modules/blogs/interfaces/services/create.blog.service.interface.ts
@@ -1,6 +1,0 @@
-import { CreateBlogDomain } from '../../domain/request/create.blog.domain';
-import { BlogEntity } from '../../domain/blog.entity';
-
-export interface CreateBlogService {
-    createBlog(blogData: CreateBlogDomain): Promise<BlogEntity>;
-}
