--- src/modules/blogs/interfaces/services/get.blog.service.interface.ts
@@ -1,5 +1,0 @@
-import { BlogEntity } from '../../domain/blog.entity';
-
-export interface GetBlogService {
-    getBlog(blogId: number): Promise<BlogEntity>;
-}
