--- src/modules/blogs/interfaces/types.ts
@@ -1,10 +1,0 @@
-export const TYPES = {
-    services: {
-        CreateBlogService: 'CreateBlogService',
-        GetBlogService: 'GetBlogService',
-    },
-    applications: {
-        CreateBlogApplication: 'CreateBlogApplication',
-        GetBlogApplication: 'GetBlogApplication',
-    },
-};
