+++ src/modules/bookmarks/domain/bookmark.domain.ts
@@ -1,1 +1,1 @@
-// domain
+// entities
