+++ src/modules/calendar/domain/calendar.domain.ts
@@ -1,1 +1,1 @@
-// domain
+// entities
