+++ src/modules/conferences/domain/conference.domain.ts
@@ -1,1 +1,1 @@
-// domain
+// entities
