+++ src/modules/events/domain/event.domain.ts
@@ -1,1 +1,1 @@
-// domain
+// entities
