+++ src/modules/jobinfo/domain/jobinfo.domain.ts
@@ -1,1 +1,1 @@
-// domain
+// entities
