+++ src/modules/resumes/domain/resume.domain.ts
@@ -1,1 +1,1 @@
-// domain
+// entities
