+++ src/modules/teams/domain/team.domain.ts
@@ -1,1 +1,1 @@
-// domain
+// entities
